dashboard/requirements.txt
flask==3.1.0
flask-socketio==5.5.1
pyserial==3.5
